// src/config/swagger/swagger.constants.ts
export enum Tags {
  HEALTH = 'Health',
  AUTHENTICATION = 'Authentication 🔐',
  PERIODS = 'Periods 🗓️',
  PERIOD_GROUPS = 'Period Groups 🅿️👥',
  GROUPS = 'Groups 👥',
  PERIOD_TIMESLOTS = 'Period timeslots 🗓️⌚',
  TIME_SLOTS = 'Time Slots ⌚',
  STUDENTS = 'Students 👦',
  TEACHERS = 'Teachers 👩‍🏫',
  GROUP_STUDENTS = 'Group Students 👥🎒',
  ENROLLMENTS = 'Enrollments 🧑📚',
  STUDENT_ENROLLMENTS = 'Student enrollments 👦📚',
  SUBJECTS = 'Subjects 📚',
  SUBJECT_GROUPS = 'Subject Groups 📚👥',
  SUBJECT_GROUP_TIME_SLOTS = 'SubjectGroupTimeSlots ⌚📚',
  ABSENCES = 'Absences 🚨',
  PERMISSIONS = 'Permissions 🅿️',
  PERMISSION_REQUESTS = 'Permission Requests 🅿️🙋‍♂️',
  DAILY_REPORTS = 'Daily Reports 📃',
  SCHEDULES = 'Schedules 📜',
  USERS = 'Users 👤',
  STATS = 'Stats 📊',
}